+++ domain/build.gradle.kts
@@ -34,9 +34,18 @@ android {
 
 dependencies {
 
+    implementation(project(mapOf("path" to ":utilities")))
     implementation("androidx.core:core-ktx:1.7.0")
     implementation("androidx.appcompat:appcompat:1.4.2")
     implementation("com.google.android.material:material:1.6.1")
+
+    /** mesh network **/
+    implementation("live.ditto:ditto:2.0.0-alpha1")
+
+    /** reactive stream coroutines **/
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1")
+
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.3")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.4.0")
